+++ photoframe/asgi.py
@@ -1,4 +1,4 @@
-"""ASGI config for the Photo Frame 2D project."""
+"""ASGI config for the Photo Frame 3D project."""
 
 import os